Addition of archive file for unused methods
