+++ Model/backend.py
@@ -561,131 +561,6 @@ def apply_forces_gpu(cell_jkr_force, cell_motility_force, cell_locations, cell_r
                 cell_locations[index][i] = new_location
 
 
-@jit(nopython=True, parallel=True)
-def nearest_cpu(number_cells, cell_locations, bins, bins_help, distance, if_diff, gata6_high, nanog_high, nearest_gata6,
-                nearest_nanog, nearest_diff):
-    """ this is the just-in-time compiled version of nearest
-        that runs in parallel on the cpu
-    """
-    # loops over all cells, with the current cell index being the focus
-    for focus in prange(number_cells):
-        # offset bins by 2 to avoid missing cells that fall outside the space
-        block_location = cell_locations[focus] // distance + np.array([2, 2, 2])
-        x, y, z = int(block_location[0]), int(block_location[1]), int(block_location[2])
-
-        # initialize these variables with essentially nothing values and the distance as an initial comparison
-        nearest_gata6_index, nearest_nanog_index, nearest_diff_index = np.nan, np.nan, np.nan
-        nearest_gata6_dist, nearest_nanog_dist, nearest_diff_dist = distance * 2, distance * 2, distance * 2
-
-        # loop over the bin the cell is in and the surrounding bin
-        for i in range(-1, 2):
-            for j in range(-1, 2):
-                for k in range(-1, 2):
-                    # get the count of cells for the current bin
-                    bin_count = bins_help[x + i][y + j][z + k]
-
-                    # go through that bin
-                    for l in range(bin_count):
-                        # get the index of the current cell in question
-                        current = bins[x + i][y + j][z + k][l]
-
-                        # check to see if that cell is within the search radius and not the same cell
-                        mag = np.linalg.norm(cell_locations[current] - cell_locations[focus])
-                        if mag <= distance and focus != current:
-                            # update the nearest differentiated cell first
-                            if if_diff[current]:
-                                # if it's closer than the last cell, update the nearest magnitude and index
-                                if mag < nearest_diff_dist:
-                                    nearest_diff_index = current
-                                    nearest_diff_dist = mag
-
-                            # update the nearest gata6 high cell making sure not nanog high
-                            elif gata6_high[current]:
-                                if not nanog_high[current]:
-                                    # if it's closer than the last cell, update the nearest magnitude and index
-                                    if mag < nearest_gata6_dist:
-                                        nearest_gata6_index = current
-                                        nearest_gata6_dist = mag
-
-                            # update the nearest nanog high cell
-                            elif nanog_high[current]:
-                                # if it's closer than the last cell, update the nearest magnitude and index
-                                if mag < nearest_nanog_dist:
-                                    nearest_nanog_index = current
-                                    nearest_nanog_dist = mag
-
-        # update the nearest cell of desired type
-        nearest_gata6[focus] = nearest_gata6_index
-        nearest_nanog[focus] = nearest_nanog_index
-        nearest_diff[focus] = nearest_diff_index
-
-    # return the updated edges
-    return nearest_gata6, nearest_nanog, nearest_diff
-
-
-@cuda.jit
-def nearest_gpu(cell_locations, bins, bins_help, distance, if_diff, gata6_high, nanog_high, nearest_gata6,
-                nearest_nanog, nearest_diff):
-    """ this is the cuda kernel for the nearest function
-        that runs on a NVIDIA gpu
-    """
-    # get the index in the array
-    focus = cuda.grid(1)
-
-    # checks to see that position is in the array
-    if focus < cell_locations.shape[0]:
-        # offset bins by 2 to avoid missing cells that fall outside the space
-        x = int(cell_locations[focus][0] / distance[0]) + 2
-        y = int(cell_locations[focus][1] / distance[0]) + 2
-        z = int(cell_locations[focus][2] / distance[0]) + 2
-
-        # initialize these variables with essentially nothing values and the distance as an initial comparison
-        nearest_gata6_index, nearest_nanog_index, nearest_diff_index = np.nan, np.nan, np.nan
-        nearest_gata6_dist, nearest_nanog_dist, nearest_diff_dist = distance[0] * 2, distance[0] * 2, distance[0] * 2
-
-        # loop over the bin the cell is in and the surrounding bins
-        for i in range(-1, 2):
-            for j in range(-1, 2):
-                for k in range(-1, 2):
-                    # get the count of cells for the current bin
-                    bin_count = bins_help[x + i][y + j][z + k]
-
-                    # go through that bin determining if a cell is a neighbor
-                    for l in range(bin_count):
-                        # get the index of the current cell in question
-                        current = bins[x + i][y + j][z + k][l]
-
-                        # check to see if that cell is within the search radius and not the same cell
-                        mag = magnitude(cell_locations[focus], cell_locations[current])
-                        if mag <= distance[0] and focus != current:
-                            # update the nearest differentiated cell first
-                            if if_diff[current]:
-                                # if it's closer than the last cell, update the nearest magnitude and index
-                                if mag < nearest_diff_dist:
-                                    nearest_diff_index = current
-                                    nearest_diff_dist = mag
-
-                            # update the nearest gata6 high cell making sure not nanog high
-                            elif gata6_high[current]:
-                                if not nanog_high[current]:
-                                    # if it's closer than the last cell, update the nearest magnitude and index
-                                    if mag < nearest_gata6_dist:
-                                        nearest_gata6_index = current
-                                        nearest_gata6_dist = mag
-
-                            # update the nearest nanog high cell
-                            elif nanog_high[current]:
-                                # if it's closer than the last cell, update the nearest magnitude and index
-                                if mag < nearest_nanog_dist:
-                                    nearest_nanog_index = current
-                                    nearest_nanog_dist = mag
-
-        # update the nearest cell of certain types
-        nearest_gata6[focus] = nearest_gata6_index
-        nearest_nanog[focus] = nearest_nanog_index
-        nearest_diff[focus] = nearest_diff_index
-
-
 @jit(nopython=True)
 def setup_diffuse_bins_cpu(diffuse_locations, diffuse_radius, diffuse_bins, diffuse_bins_help):
     """ this is the just-in-time compiled version of
@@ -855,113 +730,6 @@ def highest_fgf4_gpu(cell_locations, diffuse_bins, diffuse_bins_help, diffuse_lo
         cell_highest_fgf4[focus][0] = highest_index_x
         cell_highest_fgf4[focus][1] = highest_index_y
         cell_highest_fgf4[focus][2] = highest_index_z
-
-
-@jit(nopython=True, parallel=True)
-def remove_gata6_edges(length, fds, edges, delete, delete_help):
-    """ used by the outside cluster function to
-        remove differentiated/gata6 high edges.
-    """
-    # go through edges
-    for i in prange(length):
-        # add to the delete array if either node is differentiated/gata6 high
-        if fds[edges[i][0]][2] or fds[edges[i][1]][2]:
-            delete[i] = i
-            delete_help[i] = 1
-
-    return delete, delete_help
-
-
-@jit(nopython=True, parallel=True)
-def nearest_cluster_cpu(number_cells, cell_locations, bins, bins_help, distance, if_nanog, cell_cluster_nearest,
-                        members):
-    """ This is the Numba optimized version of
-        the nearest_cluster function.
-    """
-    # loops over all cells, with the current cell index being the focus
-    for focus in prange(number_cells):
-        if if_nanog[focus]:
-            # offset bins by 2 to avoid missing cells that fall outside the space
-            block_location = cell_locations[focus] // distance + np.array([2, 2, 2])
-            x, y, z = int(block_location[0]), int(block_location[1]), int(block_location[2])
-
-            # initialize these variables with essentially nothing values and the distance as an initial comparison
-            nearest_cell_index = np.nan
-            nearest_cell_dist = distance * 2
-
-            # loop over the bin the cell is in and the surrounding bin
-            for i in range(-1, 2):
-                for j in range(-1, 2):
-                    for k in range(-1, 2):
-                        # get the count of cells for the current bin
-                        bin_count = bins_help[x + i][y + j][z + k]
-
-                        # go through that bin
-                        for l in range(bin_count):
-                            # get the index of the current cell in question
-                            current = bins[x + i][y + j][z + k][l]
-
-                            # make sure not differentiated, not same cell, and not in the same cluster
-                            if if_nanog[current] and focus != current and members[focus] != members[current]:
-                                # check to see if that cell is within the search radius and not the same cell
-                                mag = np.linalg.norm(cell_locations[current] - cell_locations[focus])
-                                if mag <= distance:
-                                    # if it's closer than the last cell, update the nearest magnitude and index
-                                    if mag < nearest_cell_dist:
-                                        nearest_cell_index = current
-                                        nearest_cell_dist = mag
-
-            # update the nearest cell of desired type
-            cell_cluster_nearest[focus] = nearest_cell_index
-
-    # return the updated edges
-    return cell_cluster_nearest
-
-
-@cuda.jit
-def nearest_cluster_gpu(cell_locations, bins, bins_help, distance, if_nanog, cell_nearest_cluster, members):
-    """ this is the cuda kernel for the nearest function
-        that runs on a NVIDIA gpu
-    """
-    # get the index in the array
-    focus = cuda.grid(1)
-
-    # checks to see that position is in the array
-    if focus < cell_locations.shape[0]:
-        if if_nanog[focus]:
-            # offset bins by 2 to avoid missing cells that fall outside the space
-            x = int(cell_locations[focus][0] / distance[0]) + 2
-            y = int(cell_locations[focus][1] / distance[0]) + 2
-            z = int(cell_locations[focus][2] / distance[0]) + 2
-
-            # initialize these variables with essentially nothing values and the distance as an initial comparison
-            nearest_cell_index = np.nan
-            nearest_cell_dist = distance[0] * 2
-
-            # loop over the bin the cell is in and the surrounding bin
-            for i in range(-1, 2):
-                for j in range(-1, 2):
-                    for k in range(-1, 2):
-                        # get the count of cells for the current bin
-                        bin_count = bins_help[x + i][y + j][z + k]
-
-                        # go through that bin
-                        for l in range(bin_count):
-                            # get the index of the current cell in question
-                            current = bins[x + i][y + j][z + k][l]
-
-                            # make sure not differentiated, not same cell, and not in the same cluster
-                            if if_nanog[current] and focus != current and members[focus] != members[current]:
-                                # check to see if that cell is within the search radius and not the same cell
-                                mag = magnitude(cell_locations[current], cell_locations[focus])
-                                if mag <= distance[0]:
-                                    # if it's closer than the last cell, update the nearest magnitude and index
-                                    if mag < nearest_cell_dist:
-                                        nearest_cell_index = current
-                                        nearest_cell_dist = mag
-
-            # update the nearest cell of desired type
-            cell_nearest_cluster[focus] = nearest_cell_index
 
 
 @cuda.jit(device=True)
